// src/model/IJsonModel.ts
export interface IJsonRect {
	x: number;
	y: number;
	width: number;
	height: number;
}

export interface IJsonPopout {
    layout: IJsonRowNode;
	rect: IJsonRect ;
}

export interface IJsonFloating {
    tabId: string;
    rect: IJsonRect;
    zIndex: number;
    originalParentId: string;
    originalIndex: number;
    isActive?: boolean;
}